backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
bcrypt==4.1.3
openpyxl==3.1.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
Pillow==10.4.0
